[MOV] b_shift: res.partner definition in separated file
