+++ beesdoo_shift/models/__init__.py
@@ -2,3 +2,4 @@
 from . import task
 from . import planning
 from . import cooperative_status
+from . import res_partner
